requirements.txt
# ============================================
# CheXQuery-MedVLM Requirements
# ============================================

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0

# Transformers & PEFT
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vision Models
timm>=0.9.12
open_clip_torch>=2.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
datasets>=2.14.0
kagglehub>=0.3.0

# Evaluation Metrics
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13
f1chexbert>=0.0.2
pycocoevalcap>=1.2

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
einops>=0.7.0

# Logging and Visualization
wandb>=0.16.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Code Quality
black>=23.0.0
isort>=5.12.0
